Add files via upload
In this sheet to generate the 12 Word Seed Phrase each of the 128 cells first chooses a random number between 1,2,3,4,5,6. If the number it picks is even it adds a binary 0 and if the number is odd it adds a 1 binary. So it simulates 128 random dice rolls to make a long 128 bit string in binary then converts each lot of 11 binary to a word and the last 7 bits to a word which indicates which block to look for the 12th Checksum word in. Also converts the 128 bits to HEX and provides external links to SHA256 to HEX result to find the first digit in the SHA256 to find the last four binary bits example 1011 to give the Checksum.