Roll Dice Offline Calculate Seed Phrases
Updated version 0.0.2
Just improved instructions to be clearer. The template itself has not changed from the first one. 
Remember to be truly offline you must use a calculator that has no WiFi ect even if it is tempted to use a mobile phone calculator do not as phone is connected online and has WiFi etc. A hacker could have access to your phone. Print this pdf on paper sit down roll your dices and figure it all out offline enter 12th or 24th word in a hardware wallet also offline. Everything is done offline and one of the safest ways around to derive a 12 or 24 word seed and as random entropy as you can get provided your dice is not bias. But little difference it will make in this instance. Can flip a coin if you want tails for 0 and 1 for heads but just try to vary the height spun from and different surfaces or such to give variety and extra randomness because you might flip a coin similar each time giving more bias on it being heads or tails each time.